requirement.txt
haystack-ai==2.0.0
datasets == 2.21.0
sentence-transformers == 3.0.1
chroma-haystack == 0.21.1
qdrant-haystack == 4.2.0
wikipedia == 1.4.0
#pip install farm-haystack